added basic files + javadoc for interfaces

+++ src/main/java/inf112/skeleton/app/model/Effect.java
@@ -1,3 +1,8 @@
 package inf112.skeleton.app.model;
 
-public record Effect(int duration, double speedBoost, double jumpBoost) {}
+/**
+ * @param duration number of steps the effect is applied
+ * @param speedBoost speed multiplier
+ * @param jumpBoost jump boost multiplier
+ */
+public record Effect(int duration, float speedBoost, float jumpBoost) {}
